backend/pyproject.toml
[project]
name = "crabgrass"
version = "0.1.0"
description = "Idea-to-innovation platform with AI agents"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "duckdb>=1.1.0",
    "google-genai>=1.0.0",
    "google-adk>=1.0.0",
    "ag-ui-adk>=0.1.0",
    "blinker>=1.9.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crabgrass"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
